+++ src/display/mask/GeometryMask.js
@@ -131,6 +131,20 @@ var GeometryMask = new Class({
     postRenderCanvas: function (renderer)
     {
         renderer.currentContext.restore();
+    },
+
+    /**
+     * Destroys this GeometryMask and nulls any references it holds.
+     * 
+     * Note that if a Game Object is currently using this mask it will _not_ automatically detect you have destroyed it,
+     * so be sure to call `clearMask` on any Game Object using it, before destroying it.
+     *
+     * @method Phaser.Display.Masks.GeometryMask#destroy
+     * @since 3.6.1
+     */
+    destroy: function ()
+    {
+        this.geometryMask = null;
     }
 
 });