+++ src/loader/filetypes/BinaryFile.js
@@ -31,21 +31,22 @@ var BinaryFile = new Class({
 
     initialize:
 
-    function BinaryFile (key, url, path, xhrSettings)
+    function BinaryFile (loader, key, url, xhrSettings)
     {
         var fileKey = (typeof key === 'string') ? key : GetFastValue(key, 'key', '');
 
         var fileConfig = {
             type: 'binary',
+            cache: loader.cacheManager.binary,
             extension: GetFastValue(key, 'extension', 'bin'),
             responseType: 'arraybuffer',
             key: fileKey,
             url: GetFastValue(key, 'file', url),
-            path: path,
+            path: loader.path,
             xhrSettings: GetFastValue(key, 'xhr', xhrSettings)
         };
 
-        File.call(this, fileConfig);
+        File.call(this, loader, fileConfig);
     },
 
     onProcess: function (callback)
@@ -85,12 +86,12 @@ FileTypesManager.register('binary', function (key, url, xhrSettings)
         for (var i = 0; i < key.length; i++)
         {
             //  If it's an array it has to be an array of Objects, so we get everything out of the 'key' object
-            this.addFile(new BinaryFile(key[i], url, this.path, xhrSettings));
+            this.addFile(new BinaryFile(this, key[i], url, xhrSettings));
         }
     }
     else
     {
-        this.addFile(new BinaryFile(key, url, this.path, xhrSettings));
+        this.addFile(new BinaryFile(this, key, url, xhrSettings));
     }
 
     //  For method chaining