+++ src/math/DegToRad.js
@@ -7,14 +7,14 @@
 var CONST = require('./const');
 
 /**
- * [description]
+ * Convert the given angle from degrees, to the equivalent angle in radians.
  *
  * @function Phaser.Math.DegToRad
  * @since 3.0.0
  *
- * @param {integer} degrees - [description]
+ * @param {integer} degrees - The angle (in degrees) to convert to radians.
  *
- * @return {float} [description]
+ * @return {float} The given angle converted to radians.
  */
 var DegToRad = function (degrees)
 {