+++ src/math/Rotate.js
@@ -5,15 +5,15 @@
  */
 
 /**
- * [description]
+ * Rotate a given point by a given angle around the origin (0, 0), in an anti-clockwise direction.
  *
  * @function Phaser.Math.Rotate
  * @since 3.0.0
  *
- * @param {(Phaser.Geom.Point|object)} point - [description]
- * @param {number} angle - [description]
+ * @param {(Phaser.Geom.Point|object)} point - The point to be rotated.
+ * @param {number} angle - The angle to be rotated by in an anticlockwise direction.
  *
- * @return {Phaser.Geom.Point} [description]
+ * @return {Phaser.Geom.Point} The given point, rotated by the given angle in an anticlockwise direction.
  */
 var Rotate = function (point, angle)
 {