+++ src/math/RoundAwayFromZero.js
@@ -5,14 +5,14 @@
  */
 
 /**
- * [description]
+ * Round a given number so it is further away from zero. That is, positive numbers are rounded up, and negative numbers are rounded down.
  *
  * @function Phaser.Math.RoundAwayFromZero
  * @since 3.0.0
  *
- * @param {number} value - [description]
+ * @param {number} value - The number to round.
  *
- * @return {number} [description]
+ * @return {number} The rounded number, rounded away from zero.
  */
 var RoundAwayFromZero = function (value)
 {