+++ src/math/Within.js
@@ -10,8 +10,8 @@
  * @function Phaser.Math.Within
  * @since 3.0.0
  *
- * @param {number} a - [description]
- * @param {number} b - [description]
+ * @param {number} a - The first value to use in the calculation.
+ * @param {number} b - The second value to use in the calculation.
  * @param {number} tolerance - The tolerance. Anything equal to or less than this value is considered as being within range.
  *
  * @return {boolean} Returns `true` if `a` is less than or equal to the tolerance of `b`.