+++ src/renderer/webgl/pipelines/components/ModelViewProjection.js
@@ -4,15 +4,47 @@
  * @license      {@link https://github.com/photonstorm/phaser/blob/master/license.txt|MIT License}
  */
 
+
+/**
+ * Implements a model view projection matrices.
+ * Pipelines can implement this for doing 2D and 3D rendering.
+ */
+
 var ModelViewProjection = {
 
+    /**
+     * Dirty flag for checking if model matrix needs to be updated on GPU.
+     */
     modelMatrixDirty: false,
+
+    /**
+     * Dirty flag for checking if view matrix needs to be updated on GPU.
+     */
     viewMatrixDirty: false,
+
+    /**
+     * Dirty flag for checking if projection matrix needs to be updated on GPU.
+     */
     projectionMatrixDirty: false,
+
+    /**
+     * Model matrix
+     */
     modelMatrix: null,
+
+    /**
+     * View matrix
+     */
     viewMatrix: null,
+
+    /**
+     * Projection matrix
+     */
     projectionMatrix: null,
 
+    /**
+     * Initializes MVP matrices with an identity matrix
+     */
     mvpInit: function ()
     {
         this.modelMatrixDirty = true;
@@ -43,6 +75,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * If dirty flags are set then the matrices are uploaded to the GPU.
+     */
     mvpUpdate: function ()
     {
         var program = this.program;
@@ -68,6 +103,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Loads an identity matrix to the model matrix
+     */
     modelIdentity: function ()
     {
         var modelMatrix = this.modelMatrix;
@@ -94,6 +132,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Scale model matrix
+     */
     modelScale: function (x, y, z)
     {
         var modelMatrix = this.modelMatrix;
@@ -116,6 +157,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Translate model matrix
+     */
     modelTranslate: function (x, y, z)
     {
         var modelMatrix = this.modelMatrix;
@@ -130,6 +174,10 @@ var ModelViewProjection = {
         return this;
     },
 
+
+    /**
+     * Rotates the model matrix in the X axis.
+     */
     modelRotateX: function (radians)
     {
         var modelMatrix = this.modelMatrix;
@@ -158,6 +206,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Rotates the model matrix in the Y axis.
+     */
     modelRotateY: function (radians)
     {
         var modelMatrix = this.modelMatrix;
@@ -185,7 +236,10 @@ var ModelViewProjection = {
         
         return this;
     },
-
+    
+    /**
+     * Rotates the model matrix in the Z axis.
+     */
     modelRotateZ: function (radians)
     {
         var modelMatrix = this.modelMatrix;
@@ -214,6 +268,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Loads identity matrix into the view matrix
+     */
     viewIdentity: function ()
     {
         var viewMatrix = this.viewMatrix;
@@ -239,7 +296,10 @@ var ModelViewProjection = {
         
         return this;
     },
-
+    
+    /**
+     * Scales view matrix
+     */
     viewScale: function (x, y, z)
     {
         var viewMatrix = this.viewMatrix;
@@ -262,6 +322,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Translates view matrix
+     */
     viewTranslate: function (x, y, z)
     {
         var viewMatrix = this.viewMatrix;
@@ -275,7 +338,10 @@ var ModelViewProjection = {
 
         return this;
     },
-
+    
+    /**
+     * Rotates view matrix in the X axis.
+     */
     viewRotateX: function (radians)
     {
         var viewMatrix = this.viewMatrix;
@@ -303,7 +369,10 @@ var ModelViewProjection = {
 
         return this;
     },
-
+    
+    /**
+     * Rotates view matrix in the Y axis.
+     */
     viewRotateY: function (radians)
     {
         var viewMatrix = this.viewMatrix;
@@ -331,7 +400,10 @@ var ModelViewProjection = {
         
         return this;
     },
-
+    
+    /**
+     * Rotates view matrix in the Z axis.
+     */
     viewRotateZ: function (radians)
     {
         var viewMatrix = this.viewMatrix;
@@ -360,6 +432,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Loads a 2D view matrix (3x2 matrix) into a 4x4 view matrix 
+     */
     viewLoad2D: function (matrix2D)
     {
         var vm = this.viewMatrix;
@@ -386,6 +461,10 @@ var ModelViewProjection = {
         return this;
     },
 
+
+    /**
+     * Copies a 4x4 matrix into the view matrix
+     */
     viewLoad: function (matrix)
     {
         var vm = this.viewMatrix;
@@ -411,7 +490,10 @@ var ModelViewProjection = {
 
         return this;
     },
-
+    
+    /**
+     * Loads identity matrix into the projection matrix.
+     */
     projIdentity: function ()
     {
         var projectionMatrix = this.projectionMatrix;
@@ -438,6 +520,9 @@ var ModelViewProjection = {
         return this;
     },
 
+    /**
+     * Sets up an orthographics projection matrix
+     */
     projOrtho: function (left, right, bottom, top, near, far)
     {
         var projectionMatrix = this.projectionMatrix;
@@ -465,7 +550,10 @@ var ModelViewProjection = {
         this.projectionMatrixDirty = true;
         return this;
     },
-
+    
+    /**
+     * Sets up a perspective projection matrix
+     */
     projPersp: function (fovy, aspectRatio, near, far)
     {
         var projectionMatrix = this.projectionMatrix;