+++ src/textures/TextureSource.js
@@ -4,8 +4,8 @@
  * @license      {@link https://github.com/photonstorm/phaser/blob/master/license.txt|MIT License}
  */
 
+var CanvasPool = require('../display/canvas/CanvasPool');
 var Class = require('../utils/Class');
-var CONST = require('../const');
 var IsSizePowerOfTwo = require('../math/pow2/IsSizePowerOfTwo');
 var ScaleModes = require('../renderer/ScaleModes');
 
@@ -33,6 +33,15 @@ var TextureSource = new Class({
     function TextureSource (texture, source, width, height)
     {
         var game = texture.manager.game;
+
+        /**
+         * The Texture this TextureSource belongs to.
+         *
+         * @name Phaser.Textures.TextureSource#renderer
+         * @type {(Phaser.Renderer.Canvas.CanvasRenderer|Phaser.Renderer.WebGL.WebGLRenderer)}
+         * @since 3.6.1
+         */
+        this.renderer = game.renderer;
 
         /**
          * The Texture this TextureSource belongs to.
@@ -143,9 +152,16 @@ var TextureSource = new Class({
      */
     init: function (game)
     {
-        if (game.config.renderType === CONST.WEBGL)
+        if (this.renderer.gl)
         {
-            this.glTexture = game.renderer.createTextureFromSource(this.image, this.width, this.height, this.scaleMode);
+            if (this.isCanvas)
+            {
+                this.glTexture = this.renderer.canvasToTexture(this.image);
+            }
+            else
+            {
+                this.glTexture = this.renderer.createTextureFromSource(this.image, this.width, this.height, this.scaleMode);
+            }
         }
 
         if (game.config.pixelArt)
@@ -168,24 +184,47 @@ var TextureSource = new Class({
      */
     setFilter: function (filterMode)
     {
-        var game = this.texture.manager.game;
-
-        if (game.config.renderType === CONST.WEBGL)
+        if (this.renderer.gl)
         {
-            game.renderer.setTextureFilter(this.glTexture, filterMode);
+            this.renderer.setTextureFilter(this.glTexture, filterMode);
         }
     },
 
     /**
-     * Destroys this Texture Source and nulls the source image reference.
+     * If this TextureSource is backed by a Canvas and is running under WebGL,
+     * it updates the WebGLTexture using the canvas data.
+     *
+     * @method Phaser.Textures.TextureSource#update
+     * @since 3.6.1
+     */
+    update: function ()
+    {
+        if (this.renderer.gl && this.isCanvas)
+        {
+            this.renderer.canvasToTexture(this.image, this.glTexture);
+        }
+    },
+
+    /**
+     * Destroys this Texture Source and nulls the references.
      *
      * @method Phaser.Textures.TextureSource#destroy
      * @since 3.0.0
      */
     destroy: function ()
     {
-        this.texture = null;
+        if (this.glTexture)
+        {
+            this.renderer.deleteTexture(this.glTexture);
+        }
 
+        if (this.isCanvas)
+        {
+            CanvasPool.remove(this.image);
+        }
+
+        this.renderer = null;
+        this.texture = null;
         this.image = null;
     }